// apps/fusion/project/ws/app/src/lib/routes/search/models/search.model.ts
export interface ISearchRequest {
  filters: {
    [key: string]: string[];
  }
  advancedFilters?: {
    title: string;
    filters: {
      [type: string]: string[];
    };
  }[]
  query: string
  isStandAlone?: boolean
  instanceCatalog?: boolean
  pageNo?: number
  pageSize?: number
  // sortBy?: 'lastUpdatedOn'
  // sortOrder?: 'ASC' | 'DESC'
  sort?: { [sortType: string]: string }[]
  locale?: string[]
}
// ...
export interface IFeatureSearchConfig {
  tabs: ISearchConfigContentStrip[]
  routeValue: string[]
  placeHolder: { [key: string]: string }
  social: {
    qanda?: {
      latest: { dtLastModified: string };
      trending: { upVoteCount: string };
    };
    blogs?: {
      latest: {
        dtLastModified: string;
      };
      trending: {
        likeCount: string;
      };
    };
  }
}

export interface INsSearchRedirection {
  f?: {
    [index: string]: string[];
  }
  q?: string
  tab?: string
}

export interface ISearchConfigContentStrip {
  titleKey?: string
  title?: string
  reqRoles?: string[]
  reqFeatures?: string[]
  searchRedirection?: INsSearchRedirection
  searchQuery?: ISearchRequest
  contentIds?: string[]
}